nuget folder may not exist

+++ Build/scripts/Clear-NugetCache.ps1
@@ -2,5 +2,7 @@ function Clear-NugetCache {
     param ()
 
     $nugetPackages = Get-NugetPath
-    Get-ChildItem -Path $nugetPackages -Filter 'servicemodel.grpc*' -Directory | Remove-Item -Force -Recurse
+    if (Test-Path $nugetPackages) {
+        Get-ChildItem -Path $nugetPackages -Filter 'servicemodel.grpc*' -Directory | Remove-Item -Force -Recurse
+    }
 }
